RNG: replace GSL with Intel's scalar RNG
This is step 1 in a 2-step process to vectorize the RNG. Intel has a vector
RNG, which we will implement in step 2.  Both the scalar and vector RNGs are
from MKL and so are not portable. With some strategic #ifdef's we can maybe do
something similar if other compiler vendors have vector RNGs.